LPAR-08: fix shaped-text test regressions and clear clippy --all-targets
Two classes of cleanup on the post-GPT LPAR-08 tree:

1. Test regressions from the meter/list -> shaped-text migration. Three
   committed tests were red because their renderers override draw_text but
   not draw_text_shaped, so migrated widgets fell through to the default
   extent-visualizer (which blits glyph boxes via blend_rect):
   - audio_meters_integration: Counter now counts draw_text_shaped, so
     "drew text" assertions hold on the shaped path.
   - golden_list / list_draw_selected: DisplayRenderer no-ops
     draw_text_shaped (these check background/highlight, not text), fixing
     a wrong-pixels assertion and an out-of-bounds flush panic.

2. clippy --all-targets -D warnings now clean across core/widgets/ui/
   platform (mechanical: same-type casts, redundant matching, manual
   Range::contains, useless conversions, unused import/mut, slice fill,
   map_or->is_none_or, identity/erasing-op in test indices). ui/examples/
   demo.rs gets #![allow(deprecated)] since it intentionally exercises the
   LPAR-07-deprecated ui style surface.

Gates: cargo fmt clean; clippy --all-targets -D warnings clean; per-crate
tests green (core 290, widgets 137, ui 41, platform 169).

+++ core/tests/lpar16_conformance.rs
@@ -117,7 +117,7 @@ impl Renderer for FrameBuffer {
             if cov == 0 {
                 continue;
             }
-            let alpha = ((src_alpha * u16::from(cov)) / 255) as u16;
+            let alpha = (src_alpha * u16::from(cov)) / 255;
             if alpha == 0 {
                 continue;
             }
